its not good to store coordinates as integers. Fixed it :) (We're functional baby)

+++ server/search/Classes/Location.py
@@ -5,16 +5,16 @@ class Location:
         '''Takes x and y longitude and latitude'''
         self._coords = [None,None]
         if x and y:
-            self._coords = [int(x),int(y)]
+            self._coords = [float(x), float(y)]
     def getLocation(self):
         '''Returns coordinates'''
         return [self.getX(),self.getY()]
     def setX(self,x):
         '''Sets x coordinate'''
-        self._coords[0]=int(x)
+        self._coords[0]=float(x)
     def setY(self,y):
         '''Sets y coordinate'''
-        self._coords[1]=int(y)
+        self._coords[1]=float(y)
     def getX(self):
         '''Returns X coordinate'''
         return self._coords[0]
@@ -23,8 +23,8 @@ class Location:
         return self._coords[1]
     def resetCoords(self, x, y):
         '''Sets new coordinates'''
-        self.setX(int(x))
-        self.setY(int(y))
+        self.setX(float(x))
+        self.setY(float(y))
     def distance(self, loc):
         '''Returns float distance between two locations'''
         return float(math.sqrt(((self.getX()-loc.getX())**2)+((self.getY()-loc.getY())**2)))
